L1-P1-Serialize/Deserialize-Functions
I've added the necessary code for the Serialize and Deserialize functions that satisfies all test cases. Took me some time to refresh my memory on these functions and having to pass all test cases, but I managed to finish it in 2 hours at least.

I've also added comments to make my code readable. I hope I could make it shorter and more efficient, but that is the best I can do for now.

Would be very willing and open for criticism and improvements if needed. If you need me to explain the code as well, I can do so too.