L1-P2-ObjecID-UUID-Creation
I had a hard time satisfying the test scripts given o this problem and took me quiet a few refactors before I got it right. I had difficulty creating the logic for the different timestamps buffer and the lexicographic ordering. But after a few references and stackoverflow threads, I was manage to create an ObjectID(UUID) based on what was asked. As seen on the code, I declared 3 static variables for for a random number (as given in the problem), a counter variable that acts as an incremental and initialization variable for the timestamp and the last 3 bytes, and a prevTimestamp declared as BigInt for checking on timestamp and giving a unique buffer. Overall, it took me 2+ hours in getting all test scripts passed, but I managed to. do so.