L2-P1-Execution-Cache
The problem is very straightforward. In my code, i added maps for execution storage: cache for completed, and queue for ongoing. I made sure every checks are made and return the necessary execution based on the test scripts. This problem didn't take me a lot of time, and was able to finish it in less than an hour. Comments are added for more clarity and visibility of the code.