L2-P2-Downime-Logs
This last one had me thinking quiet some time. I've seen a lot of references in stackoverflow but some codes doesn't satisfy with the test script presented (the overlaps). I remembered doing a downtime log in a project i've worked on before but was only pulled from one resource. I applied the same concept, but this time merging the logs into a single set and each set as a list of tuples (as mentioned in the problem) where a tuple represents a downtime period. I added comments to all lines so the code is readable and explained properly. This took me a lot of research and coded it for more than 3+ hours to finish.

Thank you!